remove all the async stuff and the ability to record from gui
