+++ src/recording/mod.rs
@@ -1,10 +1,19 @@
+pub mod audio_recorder;
 pub mod dbus;
 pub mod dbus_event;
-mod recorder;
-mod recording_status;
-pub mod recording_thread;
-mod recording_thread_handle;
-pub mod recording_thread_handle_status;
+pub mod recorder;
+pub use audio_recorder::AudioRecorder;
+pub use audio_recorder::SoundServer;
 
-pub use recorder::Recorder;
-pub use recorder::SoundServer;
+#[derive(PartialEq)]
+pub enum RecordingStatus {
+    FinishedOrInterrupted,
+    AlbumFinished,
+    NoNewSongForTooLong,
+}
+
+pub enum RecorderStatus {
+    Running,
+    Failed(anyhow::Error),
+    Stopped,
+}
