--- src/recording/recording_status.rs
@@ -1,6 +1,0 @@
-#[derive(PartialEq)]
-pub enum RecordingStatus {
-    FinishedOrInterrupted,
-    AlbumFinished,
-    NoNewSongForTooLong,
-}
